pyproject.toml
[project]
name = "autotrd"
version = "0.1.0"
description = "Automated US stock swing trading system"
requires-python = ">=3.9"
dependencies = [
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "psycopg2-binary>=2.9",
    "yfinance>=0.2",
    "moomoo-api>=1.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "pandas>=2.0",
    "numpy>=1.24",
    "requests>=2.31",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.pytest.ini_options]
testpaths = ["tests"]
